Working on distance-and-detection

+++ distance-and-detection/det_dist.cpp
@@ -67,7 +67,7 @@ cv::Point3d getCentroid(const cv::Mat &depthMat, const rs2::depth_frame& dframe,
     // get center coordinates in the frame
     double frame_x_index = drect.x + drect.width/2;
     double frame_y_index = drect.y + drect.height/2;
-    double z = depthMat.at(cv::Point(frame_x_index, frame_y_index)) * depth_scale;
+    double z = depthMat.at(cv::Point((int)frame_x_index, (int)frame_y_index)) * depth_scale;
 
     // calculate x angle
     double frame_x = frame_x_index - (depthMat.cols/2);
